Add custom UART0 driver
This UART0 driver removes most of the synchronization code. It is caller's
responsibility to ensure synchronization. What is more, it uses lightweight
FreeRTOS task notification system instead of semaphores.

Also, it is zero-copy and does most of the work inside the interrupt handler.
For this to work, the receive and transmit tasks must be realtime and fast
enough to keep the driver supplied with fresh buffers.